# feat(node-http-handler): improve stream collection performance