--- a/.changeset/fluffy-crabs-watch.md
+++ b/.changeset/fluffy-crabs-watch.md
@@ -1,0 +1,6 @@
+---
+"@smithy/fetch-http-handler": patch
+"@smithy/node-http-handler": patch
+---
+
+improve stream collection speed