--- a/packages/node-http-handler/src/stream-collector/index.ts
+++ b/packages/node-http-handler/src/stream-collector/index.ts
@@ -35,18 +35,15 @@ const isReadableStreamInstance = (stream: unknown): stream is IReadableStream =>
   typeof ReadableStream === "function" && stream instanceof ReadableStream;
 
 async function collectReadableStream(stream: IReadableStream): Promise<Uint8Array> {
-  let res = new Uint8Array(0);
+  const buffer = [];
   const reader = stream.getReader();
   let isDone = false;
   while (!isDone) {
     const { done, value } = await reader.read();
     if (value) {
-      const prior = res;
-      res = new Uint8Array(prior.length + value.length);
-      res.set(prior);
-      res.set(value, prior.length);
+      buffer.push(...value);
     }
     isDone = done;
   }
-  return res;
+  return new Uint8Array(buffer);
 }