I noticed stream collectors were doing iffy copy operations during the stream -> Uint8Array process.

```js
  /**
   * Perf results, Node.js 18
   * 
   * 100K items: 5466ms vs 3930ms
   * 10K items: 54ms vs 15ms
   * 1K items: 4.6ms vs 1.3ms
   * 100 items: 0.94ms vs 0.28ms
   * 10 items: 0.71ms vs 0.18ms
   * 1 item: 0.53ms vs 0.06ms
   */
```