@kuhe 2024-05-09
doing some more testing of practical scenarios

the existing algorithm pulls ahead if chunk size is very large like 100k, because there are relatively few Uint8Array copies in that case