@kuhe 2024-05-10
Note: this PR's implementation increases memory usage temporarily when collecting a stream.